// ReverseRaffle.Web/ClientApp/app/models/core/keycodes.ts
export enum KeyCodes {
    Zero = 48, One, Two, Three, Four, Five, Six, Seven, Eight, Nine, 
    NumZero = 96, NumOne, NumTwo, NumThree, NumFour, NumFive, NumSix, NumSeven, NumEight, NumNine,
    A = 65, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z,
    F1 = 112, F2, F3, F4, F5, F6, F7,
    Tab = 9, Enter = 13, Shift = 16, Ctrl, Alt, Esc = 27, Insert = 45
}